[SYSTEMDS-2907] Fix memory estimates dense and sparse matrices
The dense and sparse memory estimates are used for multiple purposes
including selecting distributed operations if needed. A recent change of
these memory estimates introduced a few issues which lead overflows
estimates memory of ultra-sparse matrices (which dense estimate exceed
int max, and in some cases long max).

* Fix int overflows on estimating memory of dense and spare matrices
* Fix incorrect estimates used by the compiler (header in wrong place)
* Fix missing long overflow checks

+++ src/test/java/org/apache/sysds/test/component/misc/MemoryEstimateTest.java
@@ -80,11 +80,11 @@ public class MemoryEstimateTest {
 				break;
 			case INT:
 				int[] arrayInt = new int[length];
-				assertEquals(MemoryEstimates.intArrayCost(length), measure(arrayInt));
+				assertEquals((long)MemoryEstimates.intArrayCost(length), measure(arrayInt));
 				break;
 			case DOUBLE:
 				double[] arrayDouble = new double[length];
-				assertEquals(MemoryEstimates.doubleArrayCost(length), measure(arrayDouble));
+				assertEquals((long)MemoryEstimates.doubleArrayCost(length), measure(arrayDouble));
 				break;
 			default:
 				System.out.println(arrayToMeasure.getClass().getSimpleName());
